Add mongodb in docker compose dev and improve configuration (#55)
The main change with the configuration is the automatic creaction of the
configuration json file if it does not exists.  The content of this file
will be the configuration calculated using command line arguments and
the environment.  The fallback settings is removed from the settingDB as
it is not needed anymore.

The change was made to make the configuration more intuitive and
facilitate running the tests.