Improve getting started for users (#340)
* handbook link

The handbook button displayed after the registration phase of the application and the link in the help/support section of the application have been updated.

* architecture created and related images moved to this folder

* nodes folder

nodes folder created and related images moved to this folder

* handbook images

images renamed

* Update handbook.md

* hamdbook img moved to media folder

There are some photos that I couldn't categorize, we can categorize them later if needed.

* Img file location fixed

* sidebar made more accessible

* about and sidebar

* Footer added

Mail, address and github